シリアルポートは1バイトなので、型を修正

+++ ELF_headers/kernel/kernel.c
@@ -1,6 +1,6 @@
 #include <stdint.h>
 
-void serialport_output(uint16_t ascii_code) {
+void serialport_output(uint8_t ascii_code) {
   __asm__ volatile(".intel_syntax noprefix\n"
                    "mov dx, 0x3f8\n"
                    /* "mov al, 1\n" */
@@ -14,8 +14,8 @@ void hlt() {
 }
 
 void kernel_main() {
-  uint16_t i;
-  uint16_t output_data[14] = {75, 69, 82, 78, 69, 76, 95, 83, 85, 67, 67, 69, 83, 83};
+  uint8_t i;
+  uint8_t output_data[14] = {75, 69, 82, 78, 69, 76, 95, 83, 85, 67, 67, 69, 83, 83};
 
   for (i = 0; i < 14; i++){
 	serialport_output(output_data[i]);
